pyproject.toml
[project]
name = "digest-agent-jimmyliao"
version = "0.1.0"
description = "AI-powered news digest agent with Streamlit dashboard"
requires-python = ">=3.11"
authors = [
    { name = "JimmyLiao", email = "jimmyliao@leapdesign.ai" },
]
dependencies = [
    "streamlit>=1.35.0",
    "google-genai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "feedparser>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
